feat(matching): exclude candidates with active requests from discover
Extract pure scoring helpers to matching-utils.ts, add buildExcludedUserIds.
Filter pending/accepted requests bidirectionally in discover query.

Scenario: Candidate the Student has requested is excluded from suggestions
Scenario: Candidate who has sent a request to the Student is excluded
Scenario: Declined request does not exclude candidate from suggestions
Task: #125
